# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "bookkeeping-app API Documentation",
        "description": "bookkeeping-app API Documentation",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://bookkeeping-app.chadwick.localhost",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/api/records": {
            "get": {
                "tags": [
                    "交易紀錄"
                ],
                "summary": "取得紀錄列表",
                "description": "取得紀錄列表",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "取得成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "item": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "datetime": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    },
                                                    "in": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "out": {
                                                        "type": "integer",
                                                        "example": 100
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "取得失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "取得失敗"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "交易紀錄"
                ],
                "summary": "新增紀錄",
                "description": "新增紀錄",
                "operationId": "create",
                "requestBody": {
                    "description": "新增紀錄",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item",
                                    "datetime",
                                    "in",
                                    "out"
                                ],
                                "properties": {
                                    "item": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "datetime": {
                                        "type": "string",
                                        "example": "2021-01-01 00:00:00"
                                    },
                                    "in": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "out": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "新增成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "新增失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "新增失敗"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}